requirements.txt
# Core Framework
flask==3.0.2
websockets==12.0
sqlalchemy==2.0.25
psutil==5.9.7
python-dotenv==1.0.0

# Security & Validation
bandit==1.7.5
python-magic==0.4.27
schema==0.7.5

# API Clients
openai==1.12.0
anthropic==0.19.1
google-generativeai==0.3.2

# Sandbox Management
virtualenv==20.25.0
chardet==5.2.0

# Mobile Utilities
termux-api==0.4.23
android-net==1.2.1

# Development Tools
black==24.2.0
pytest==7.4.4
